pyproject.toml
[tool.poetry]
name = "tpg26x"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
click = "^8.0.3"
pyserial = "^3.5"

[tool.poetry.dev-dependencies]

[tool.poerty.scripts]
tpg261 = "tpg261.cli:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
